etl_pipeline/requirements.txt
pandas
tqdm
numpy
requests
pytest
bs4
pyyaml
